chore: restructure directories and remove duplicated fonts
