+++ example/lib/pages/theme_page/theme_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:yaru/theme.dart';
+import 'package:yaru/yaru.dart';
 
 import 'home.dart';
