+++ lib/icons.dart
@@ -1,7 +1,7 @@
 // Animated Icons
-export 'src/icons_widgets/yaru_animated_icon.dart';
-export 'src/icons_widgets/yaru_animated_icons.dart';
+export 'src/icons/icons_widgets/yaru_animated_icon.dart';
+export 'src/icons/icons_widgets/yaru_animated_icons.dart';
 // Widgets
-export 'src/icons_widgets/yaru_placeholder_icon.dart';
-export 'src/yaru_freedesktop_icons.dart';
-export 'src/yaru_icons.dart';
+export 'src/icons/icons_widgets/yaru_placeholder_icon.dart';
+export 'src/icons/yaru_freedesktop_icons.dart';
+export 'src/icons/yaru_icons.dart';
