+++ lib/settings.dart
@@ -1,0 +1,4 @@
+export 'src/settings/gtk_constants.dart';
+export 'src/settings/inherited_theme.dart';
+export 'src/settings/settings_service.dart';
+export 'src/settings/yaru_settings.dart';
