+++ lib/src/icons/icons_widgets/icons_foundation/local_progress_mixin.dart
