+++ lib/src/icons/icons_widgets/yaru_animated_icons.dart
