+++ lib/src/icons/icons_widgets/yaru_placeholder_icon.dart
