+++ lib/src/settings/inherited_theme.dart
@@ -4,7 +4,7 @@ import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:platform_linux/platform.dart';
-import 'package:yaru/src/theme_widgets/yaru_settings.dart';
+import 'package:yaru/src/settings/yaru_settings.dart';
 
 import '../../theme.dart';
