+++ lib/src/settings/yaru_settings.dart
@@ -2,8 +2,8 @@ import 'dart:async';
 
 import 'package:flutter/foundation.dart';
 import 'package:gtk/gtk.dart';
-import 'package:yaru/src/theme_widgets/gtk_constants.dart';
-import 'package:yaru/src/theme_widgets/settings_service.dart';
+import 'package:yaru/src/settings/gtk_constants.dart';
+import 'package:yaru/src/settings/settings_service.dart';
 
 abstract class YaruSettings {
   factory YaruSettings() = YaruGtkSettings;
