--- lib/src/text/fonts/CONTRIBUTING.txt
@@ -1,21 +1,0 @@
-The Ubuntu Font Family is very long-term endeavour, and the first time
-that a professionally-designed font has been funded specifically with
-the intent of being an on-going community expanded project:
-
-  http://font.ubuntu.com/
-
-Development of the Ubuntu Font Family is undertaken on Launchpad:
-
-  http://launchpad.net/ubuntu-font-family/
-
-and this is where milestones, bug management and releases are handled.
-
-Contributions are welcomed.  Your work will be used on millions of
-computers every single day!  Following the initial bootstrapping of
-Latin, Cyrillic, Greek, Arabic and Hebrew expansion will be undertaken
-by font designers from the font design and Ubuntu communities.
-
-To ensure that the Ubuntu Font Family can be re-licensed to future
-widely-used libre font licences, copyright assignment is being required:
-
-  https://launchpad.net/~uff-contributors
