--- lib/src/text/fonts/README.txt
@@ -1,16 +1,0 @@
- ----------------------
-  Ubuntu Font Family
- ======================
-
-The Ubuntu Font Family are a set of matching new libre/open fonts in
-development during 2010--2011.  And with further expansion work and
-bug fixing during 2015.  The development is being funded by
-Canonical Ltd on behalf the wider Free Software community and the
-Ubuntu project.  The technical font design work and implementation is
-being undertaken by Dalton Maag.
-
-Both the final font Truetype/OpenType files and the design files used
-to produce the font family are distributed under an open licence and
-you are expressly encouraged to experiment, modify, share and improve.
-
-  http://font.ubuntu.com/
