--- lib/src/text/fonts/TRADEMARKS.txt
@@ -1,4 +1,0 @@
-Ubuntu and Canonical are registered trademarks of Canonical Ltd.
-
-The licence accompanying these works does not grant any rights
-under trademark law and all such rights are reserved.
