--- lib/src/text/fonts/copyright.txt
@@ -1,5 +1,0 @@
-Copyright 2010,2011 Canonical Ltd.
-
-This Font Software is licensed under the Ubuntu Font Licence, Version
-1.0.  https://launchpad.net/ubuntu-font-licence
-
