+++ lib/yaru.dart
@@ -3,5 +3,6 @@ library yaru;
 export 'constants.dart';
 export 'foundation.dart';
 export 'icons.dart';
+export 'settings.dart';
 export 'theme.dart';
 export 'widgets.dart';
